requirements.txt
pandas>=1.5.3
scikit-learn>=1.2.2
numpy>=1.24.3
matplotlib>=3.7.1
python-dateutil>=2.8.2